refactor: Output 인터페이스 사용
