--- src/main/kotlin/lotto/PickNumberPrinter.kt
@@ -1,8 +1,0 @@
-package lotto
-
-class PickNumberPrinter(private val lottoNumbers: List<LottoNumbers>) {
-    fun print() {
-        lottoNumbers.map { it.joinToString(prefix = "[", postfix = "]") }
-            .forEach { println(it) }
-    }
-}
